added orderItem validation and exceptions

+++ orderItems-service/src/main/java/org/example/orderitemsservice/OrderItemService.java
@@ -1,5 +1,6 @@
 package org.example.orderitemsservice;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -29,10 +30,16 @@ public class OrderItemService {
     }
 
     public OrderItem update(OrderItem orderItem) {
+        if(!orderItemRepository.existsById(orderItem.getId())){
+            throw new EntityNotFoundException("Order item not found");
+        }
         return orderItemRepository.save(orderItem);
     }
 
     public void delete(Long id) {
+        if(!orderItemRepository.existsById(id)){
+            throw new EntityNotFoundException("Order item not found");
+        }
         orderItemRepository.deleteById(id);
     }
 }
